requirements.txt
aniso8601==9.0.1
apsw==3.44.2.0
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
dominate==2.9.1
Flask==3.0.0
Flask-Bootstrap4==4.0.2
Flask-Moment==1.0.5
Flask-WTF==1.2.1
flask_weasyprint==1.1.0
fonttools==4.47.2
github-cli==1.0.0
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
html5lib==1.1
httplib2==0.22.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pango==0.0.1
pillow==10.2.0
platformdirs==4.1.0
protobuf==4.25.1
py-pip==0.0.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
tinycss2==1.2.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.1.0
visitor==0.1.3
weasyprint==60.2
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.2
zopfli==0.2.3
